refactor: sample 폴더 정리
